:recycle: 로컬스토리지 대결진행 식별자 비움 함수 global.js 로 이동 #301
