+++ src/main/resources/static/home/js/home.js
@@ -1,4 +1,4 @@
-import {flushPlayRecordIdsFromLocalStorage} from "../../global/js/vstopic-localstorage.js";
+import {flushPlayRecordIdsFromLocalStorage} from "../../global/global.js";
 import { HomeExceptionHandler} from "./exception/home-exception-handler.js";
 import {topicSearchParams} from "./const.js";
 import {searchTopics} from "./home-api.js";
